Changed PieceRules to a different name and involved an abstract class in the modification.
Got more comfortable with "simulating a move" (before permanently carrying out the move and drawing the board). This move simulation seems to correctly determine if a piece moving puts a FRIENDLY king in check.

Checking for fork (and maybe pin) might be natural next steps.

Eventually checkmate is a goal, of course.