With depth > 4, minimax with alpha beta pruning fills up my 16 GB RAM and takes about 20 minutes (!) to find a move after a few moves have been made and the possibilities increase.
Depth = 4 is not very strong (ELO 1200 - 1500?)

Consider using "opening books" by reading PGN files

This will certainly be convenient for automating away the need for repetitive mouse clicks.

But it is probably more useful (for engine strength) to try to improve my algorithms and data structures.

One immediate thought is to replace string positions with chars and ints, though I'm not sure how much of an improvement this will make.