Depth 5 previously took about 40 seconds. Caching dropped this to less than 10 seconds.
RAM usage also improved a bit by changing string position to ChessPosition struct (8 bytes instead of 32)

Also made AIChessGame a child class of ChessGame. This reduced the code complexity of the ChessGame class by moving the minimax (et. al) methods to the AIChessGame class (which is in a different header and cpp file to limit scrolling time).