Might be hanging it up for a while with this project.
Depth 5 is feasible now even a few moves into the game when things open up, but I think the evaluation function is responsible for making suboptimal recommendations.

Next time I come back to this, it would be good to just improve my understanding of minimax and the alpha beta pruning.